+++ Virtusize/Sources/Flutter/VirtusizeFlutterReposiory.swift
@@ -23,6 +23,7 @@
 //
 
 import Foundation
+import VirtusizeCore
 
 public class VirtusizeFlutterRepository: NSObject {
 	public static let shared: VirtusizeFlutterRepository = {
@@ -138,7 +139,7 @@ public class VirtusizeFlutterRepository: NSObject {
 		productTypes: [VirtusizeProductType],
 		storeProduct: VirtusizeServerProduct,
 		userBodyProfile: VirtusizeUserBodyProfile
-	) -> BodyProfileRecommendedSize? {
+	) -> BodyProfileRecommendedSizeArray? {
 		let response = VirtusizeAPIService.getBodyProfileRecommendedSizeAsync(
 			productTypes: productTypes,
 			storeProduct: storeProduct,