+++ Virtusize/Sources/Internal/Utils/Localization.swift
@@ -37,7 +37,7 @@ internal class Localization {
 	///   - language: Pass `VirtusizeLanguage` if you'd like to localize the text in a designated language
 	/// - Returns: A localized string based on the device's default language
 	func localize(_ key: String, language: VirtusizeLanguage? = nil) -> String {
-		return BundleLoader.localizationBundle(
+		return VirtusizeCoreBundleLoader.localizationBundle(
 			language: language?.rawValue ?? Virtusize.params?.language.rawValue
 		).localizedString(
 			forKey: key, value: nil, table: "VirtusizeLocalizable"