+++ Virtusize/Sources/Internal/Workers/Deserializer.swift
@@ -23,6 +23,7 @@
 //
 
 import Foundation
+import VirtusizeCore
 
 /// A structure that helps to deserialize the models specific to the Virtusize API
 internal struct Deserializer {