+++ Virtusize/Sources/Models/VirtusizeEnvironment.swift
@@ -24,23 +24,49 @@
 
 /// This enum contains all available Virtusize environments
 public enum VirtusizeEnvironment: String, CaseIterable {
-	case staging="staging.virtusize.jp"
-	case global="api.virtusize.com"
-	case japan="api.virtusize.jp"
-	case korea="api.virtusize.kr"
+	case TESTING="testing.virtusize.jp"
+	case STAGING="staging.virtusize.com"
+	case GLOBAL="api.virtusize.com"
+	case JAPAN="api.virtusize.jp"
+	case KOREA="api.virtusize.kr"
+
+	var isProdEnv: Bool {
+		return self == .GLOBAL || self == .JAPAN || self == .KOREA
+	}
 
 	/// Gets the services URL for the `productDataCheck` and `getSize` requests
 	internal func servicesUrl() -> String {
 		// swiftlint:disable switch_case_alignment
 		switch self {
-			case .staging, .japan:
+			case .TESTING:
 				return "services.virtusize.jp"
-			case .global:
+			case .STAGING:
 				return "services.virtusize.com"
-			case .korea:
+			case .GLOBAL:
+				return "services.virtusize.com"
+			case .JAPAN:
+				return "services.virtusize.jp"
+			case .KOREA:
 				return "services.virtusize.kr"
 		}
 	}
+    
+    /// Gets the services URL for the ``getSize` requests
+    internal func getSizeUrl() -> String {
+        // swiftlint:disable switch_case_alignment
+        switch self {
+            case .TESTING:
+                return "size-recommendation.virtusize.jp"
+            case .STAGING:
+                return "size-recommendation.virtusize.com"
+            case .GLOBAL:
+                return "size-recommendation.virtusize.com"
+            case .JAPAN:
+                return "size-recommendation.virtusize.jp"
+            case .KOREA:
+                return "size-recommendation.virtusize.kr"
+        }
+    }
 
 	/// Gets the URL for the `i18n` request
 	internal func i18nUrl() -> String {
@@ -50,11 +76,11 @@ public enum VirtusizeEnvironment: String, CaseIterable {
 	/// Gets the static API URL for the `VirtusizeWebView` request
 	internal func virtusizeStaticApiUrl() -> String {
 		switch self {
-			case .staging, .japan:
-				return "static.api.virtusize.jp"
-			case .global:
+			case .STAGING, .GLOBAL:
 				return "static.api.virtusize.com"
-			case .korea:
+			case .TESTING, .JAPAN:
+				return "static.api.virtusize.jp"
+			case .KOREA:
 				return "static.api.virtusize.kr"
 		}
 	}
@@ -62,13 +88,15 @@ public enum VirtusizeEnvironment: String, CaseIterable {
 	/// Gets the event API URL corresponding to the Virtusize environment
 	internal func eventApiUrl() -> String {
 		switch self {
-			case .staging:
-				return "events.staging.virtusize.jp"
-			case .japan:
+			case .TESTING:
+				return "events.testing.virtusize.jp"
+			case .STAGING:
+				return "events.staging.virtusize.com"
+			case .JAPAN:
 				return "events.virtusize.jp"
-			case .global:
+			case .GLOBAL:
 				return "events.virtusize.com"
-			case .korea:
+			case .KOREA:
 				return "events.virtusize.kr"
 		}
 	}
@@ -76,11 +104,11 @@ public enum VirtusizeEnvironment: String, CaseIterable {
 	/// Gets the `VirtusizeRegion` corresponding to the Virtusize environment
 	internal func virtusizeRegion() -> VirtusizeRegion {
 		switch self {
-			case .staging, .japan:
-				return VirtusizeRegion.JAPAN
-			case .global:
+			case .STAGING, .GLOBAL:
 				return VirtusizeRegion.COM
-			case .korea:
+			case .TESTING, .JAPAN:
+				return VirtusizeRegion.JAPAN
+			case .KOREA:
 				return VirtusizeRegion.KOREA
 		}
 	}