+++ Virtusize/Sources/UI/VirtusizeWebViewController.swift
@@ -24,6 +24,8 @@
 
 import UIKit
 import WebKit
+import VirtusizeCore
+import VirtusizeAuth
 
 /// The methods of this protocol notify you with Virtusize specific messages such as errors as
 /// `VirtusizeError` and events as `VirtusizeEvent`
@@ -204,6 +206,11 @@ extension VirtusizeWebViewController: WKNavigationDelegate, WKUIDelegate {
 				return nil
 			}
 		}
+
+		if VirtusizeAuthorization.isSNSAuthURL(viewController: self, webView: webView, url: url) {
+			return nil
+		}
+
 		// swiftlint:disable line_length
 		guard let targetFrame = navigationAction.targetFrame, targetFrame.isMainFrame else {
 			// By default, The Google sign-in page shows a 403 error: disallowed_useragent if you are visiting it within a Webview.