+++ VirtusizeCore/Sources/Utils/UIApplication+Extensions.swift
@@ -46,5 +46,29 @@ public extension UIApplication {
 			return
 		}
 	}
+
+    /// Gets the top most view controller
+	///
+	/// - Parameter baseVC: the base view controller
+	func getTopMostViewController() -> UIViewController? {
+       let baseVC : UIViewController? = UIApplication.safeShared?.windows.filter {$0.isKeyWindow}.first?.rootViewController
+		if let navVC = baseVC as? UINavigationController {
+			return navVC.visibleViewController
+		} else if let tabController = baseVC as? UITabBarController, let selectedTabVC = tabController.selectedViewController {
+			return selectedTabVC
+		} else if let presentedVC = baseVC?.presentedViewController {
+			return presentedVC
+		}
+		return baseVC
+	}
+}
+extension UIWindow {
+    static var key: UIWindow? {
+        if #available(iOS 13, *) {
+            return UIApplication.safeShared?.windows.first { $0.isKeyWindow }
+        } else {
+            return UIApplication.safeShared?.keyWindow
+        }
+    }
 }
 #endif