+++ VirtusizeUIKit/Sources/Utils/UIImage+Extensions.swift
@@ -23,6 +23,7 @@
 //
 
 import UIKit
+import VirtusizeCore
 
 public extension UIImage {